Move logging.js into lib folder
- it was originally outside the lib folder because it was required by some of the bin commands, now that the bin commands require the utils.js it doesn't make much sense for logging to still be outside of lib
