+++ bin/knex-migrator-health
@@ -1,35 +1,33 @@
 #!/usr/bin/env node
+'use strict';
 
-var program = require('commander');
-var utils = require('../lib/utils');
+const program = require('commander');
+const utils = require('../lib/utils');
+const logging = require('../lib/logging');
 
-var logging = require('../logging');
-var knexMigrator;
+utils.getKnexMigrator({path: process.cwd()}).then((KnexMigrator) => {
+    let knexMigrator;
 
-utils.getKnexMigrator({path: process.cwd()})
-    .then(function (KnexMigrator) {
-        program
-            .option('--mgpath <path>')
-            .parse(process.argv);
+    program
+        .option('--mgpath <path>')
+        .parse(process.argv);
 
-        try {
-            knexMigrator = new KnexMigrator({knexMigratorFilePath: program.mgpath});
-        } catch (err) {
-            logging.error(err);
-            process.exit(1);
-        }
-
-        return knexMigrator.isDatabaseOK()
-            .then(function () {
-                logging.info('Woohoo, Database is healthy');
-            });
-    })
-    .catch(function (err) {
-        logging.error(err.message);
-
-        if (err.help) {
-            logging.info(err.help);
-        }
-
+    try {
+        knexMigrator = new KnexMigrator({knexMigratorFilePath: program.mgpath});
+    } catch (err) {
+        logging.error(err);
         process.exit(1);
+    }
+
+    return knexMigrator.isDatabaseOK().then(function () {
+        logging.info('Woohoo, Database is healthy');
     });
+}).catch(function (err) {
+    logging.error(err.message);
+
+    if (err.help) {
+        logging.info(err.help);
+    }
+
+    process.exit(1);
+});
