+++ bin/knex-migrator-migrate
@@ -1,38 +1,37 @@
 #!/usr/bin/env node
+'use strict';
 
-var program = require('commander');
-var utils = require('../lib/utils');
+const program = require('commander');
+const utils = require('../lib/utils');
+const logging = require('../lib/logging');
 
-var logging = require('../logging');
-var knexMigrator;
+utils.getKnexMigrator({path: process.cwd()}).then(function (KnexMigrator) {
+    let knexMigrator;
 
-utils.getKnexMigrator({path: process.cwd()})
-    .then(function (KnexMigrator) {
-        program
-            .option('--v <item>')
-            .option('--only <item>')
-            .option('--mgpath <path>')
-            .option('--force')
-            .option('--init')
-            .parse(process.argv);
+    program
+        .option('--v <item>')
+        .option('--only <item>')
+        .option('--mgpath <path>')
+        .option('--force')
+        .option('--init')
+        .parse(process.argv);
 
-        try {
-            knexMigrator = new KnexMigrator({knexMigratorFilePath: program.mgpath});
-        } catch (err) {
-            logging.error(err);
-            process.exit(1);
-        }
-
-        return knexMigrator.migrate({
-            version: program.v,
-            only: program.only,
-            force: program.force,
-            init: program.init
-        }).then(function () {
-            logging.info('Finished database migration!');
-        });
-    })
-    .catch(function (err) {
+    try {
+        knexMigrator = new KnexMigrator({knexMigratorFilePath: program.mgpath});
+    } catch (err) {
         logging.error(err);
         process.exit(1);
+    }
+
+    return knexMigrator.migrate({
+        version: program.v,
+        only: program.only,
+        force: program.force,
+        init: program.init
+    }).then(function () {
+        logging.info('Finished database migration!');
     });
+}).catch(function (err) {
+    logging.error(err);
+    process.exit(1);
+});
