+++ bin/knex-migrator-reset
@@ -1,31 +1,29 @@
 #!/usr/bin/env node
+'use strict';
 
-var program = require('commander');
-var utils = require('../lib/utils');
+const program = require('commander');
+const utils = require('../lib/utils');
+const logging = require('../lib/logging');
 
-var logging = require('../logging');
-var knexMigrator;
+utils.getKnexMigrator({path: process.cwd()}).then(function (KnexMigrator) {
+    let knexMigrator;
 
-utils.getKnexMigrator({path: process.cwd()})
-    .then(function (KnexMigrator) {
-        program
-            .option('--mgpath <path>')
-            .option('--force')
-            .parse(process.argv);
+    program
+        .option('--mgpath <path>')
+        .option('--force')
+        .parse(process.argv);
 
-        try {
-            knexMigrator = new KnexMigrator({knexMigratorFilePath: program.mgpath});
-        } catch (err) {
-            logging.error(err);
-            process.exit(1);
-        }
-
-        return knexMigrator.reset({force: program.force})
-            .then(function () {
-                logging.info('Finished database reset!');
-            });
-    })
-    .catch(function (err) {
+    try {
+        knexMigrator = new KnexMigrator({knexMigratorFilePath: program.mgpath});
+    } catch (err) {
         logging.error(err);
         process.exit(1);
+    }
+
+    return knexMigrator.reset({force: program.force}).then(function () {
+        logging.info('Finished database reset!');
     });
+}).catch(function (err) {
+    logging.error(err);
+    process.exit(1);
+});
