+++ lib/index.js
@@ -8,7 +8,7 @@ const debug = require('ghost-ignition').debug('index');
 const database = require('./database');
 const utils = require('./utils');
 const errors = require('./errors');
-const logging = require('../logging');
+const logging = require('./logging');
 
 function KnexMigrator(options) {
     options = options || {};
