+++ lib/logging.js
@@ -1,4 +1,6 @@
-var logging = require('ghost-ignition').logging;
+'use strict';
+
+const logging = require('ghost-ignition').logging;
 
 module.exports = logging({
     env: process.env.NODE_ENV,
